requirements.txt
flask
Flask-OIDC
wsgidav
lxml
markdown
pyyaml
gunicorn
pymongo
